requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
bleach==3.0.2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
feedparser==5.2.1
http-ece==1.0.5
idna==2.8
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
MarkupSafe==1.1.0
Mastodon.py==1.3.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.2
oauthlib==2.1.0
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.2.0
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.7
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
requests-oauthlib==1.0.0
Send2Trash==1.5.0
six==1.11.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
tweepy==3.7.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
